Move non-EMMO modules to ontopy
Have absolute imports everywhere except `__init__.py` files.
