+++ emmopy/__init__.py
@@ -1,21 +1,13 @@
 # -*- coding: utf-8 -*-
 import sys
+from ontopy import __version__ as __ontopy_version__
 
-__version__ = '1.0.1'
+__version__ = __ontopy_version__
 
 # Ensure correct Python version
 if sys.version_info < (3, 6):
-    raise RuntimeError('emmo requires Python 3.6 or later')
+    raise RuntimeError('emmopy requires Python 3.6 or later')
 
-# Ensure emmo is imported before owlready2...
-if 'owlready2' in sys.modules.keys():
-    raise RuntimeError('emmo must be imported before owlready2')
-
-# Monkey patch Owlready2 by injecting some methods
-from . import patch  # noqa: E402, F401
-
-# Import World and get_ontology(), which are our main entry points
-from .ontology import World, get_ontology  # noqa: E402, F401
-
-# Global list of ontology search paths
-from owlready2 import onto_path  # noqa: E402, F401
+# Ensure emmopy is imported before owlready2...
+if 'owlready2' in sys.modules.keys() and "ontopy" not in sys.modules.keys():
+    raise RuntimeError('emmopy must be imported before owlready2')
