+++ ontopy/colortest.py
