+++ ontopy/patch.py
