+++ ontopy/utils.py
