+++ tests/test_load_emmo.py
@@ -4,7 +4,7 @@ import sys
 
 thisdir = os.path.abspath(os.path.dirname(__file__))
 sys.path.insert(1, os.path.abspath(os.path.join(thisdir, '..', '..')))
-from emmo import get_ontology  # noqa: E402, F401
+from ontopy import get_ontology  # noqa: E402, F401
 
 
 emmo = get_ontology()
