+++ tools/ontodoc
@@ -9,8 +9,8 @@ rootdir = os.path.abspath(os.path.realpath((os.path.dirname(
     os.path.dirname(__file__)))))
 sys.path.insert(1, rootdir)
 
-from emmo import World, onto_path  # noqa: E402
-from emmo.ontodoc import (
+from ontopy import World, onto_path  # noqa: E402
+from ontopy.ontodoc import (
     OntoDoc, get_format, get_style, get_figformat, get_maxwidth, get_docpp,
 )  # noqa: E402
